Refactor extension entrypoint into focused modules
Split eval flow, importer command, LSP requests, settings mapping,

and client lifecycle out of extension.ts.
